requirements.txt
mysql-connector-python
python-dotenv
pandas
openpyxl
plotly
pymysql
dash
dash-bootstrap-components